fix budget save and update ReactiveAmountField

+++ modules/feature_patient/lib/src/detail_patient/basic_info/basic_info_form.dart
@@ -159,7 +159,7 @@ FormGroup basicInfoForm({
         }),
         'MEDICAL_RECORD_BUDGETS': FormGroup({
           '_id': FormControl<String?>(),
-          'budget': FormControl<int>(
+          'budget': FormControl<double>(
             value: 0,
           ),
           'remarks': FormControl<String?>(),
